refactor: refactor DetailSkeleton, randomSizer
